+++ CHANGELOG.md
@@ -24,6 +24,11 @@
   for large files.  
   [Niil Öhlin](https://github.com/niilohlin)
 
+* Add new `legacy_hashing` rule to encourage the use of Swift 4.2's new hashing
+  interface.  
+  [Kim de Vos](https://github.com/kimdv)
+  [#2108](https://github.com/realm/SwiftLint/issues/2108)
+
 #### Bug Fixes
 
 * Fix false positives in `redundant_objc_attribute` for private declarations